chore: add hacky configuration for nightly test runs
Recently the configuration of blink changed in a way that broke the
integration-tests for the nightly version (which I run locally). This
means I cannot use them to drive development, because ci also needs to
use the stable version.

Work around this by adding a hacky configuration for the nightly
version. I think if it's being used in the stable version, it should
just get ignored.

This should not stay here for a long time - I think I'll be able to
remove it when the next stable version of blink is released.