perf: only register each unique match once (#46)
Right now we are registering each match multiple times, which is not
very useful. It will add more context to the context window
(documentation), but it's not very clear to read the context anyway.
Most of the time I just ignore it.

If in the future we have a better way to inspect each submatch, this
should be reconsidered.