perf: only collect the necessary context preview for each match
Issue
=====

The current implementation of the ripgrep parser includes all the lines
that matched in each file, and displays the same lines for each match.
As a user, I'm interested in the context around the match, not the entire
file. This is especially important when the file is large, and the file
contains a lot of matches.

Lots of unnecessary work is being done in this case.

Solution
========

Only include the necessary context preview for each match. This will
reduce the memory usage and hopefully make the plugin faster.

The speed could still be better in huge repos, but this is a step in the
right direction.