feat: fall back to regex highlighting when treesitter not available
When a result is found for a file whose filetype does not have a
treesitter parser installed, fall back to regex based highlighting that
is bundled in Neovim. This allows highlighting many more file types out
of the box.

See here for the blink side implementation and a mini demo:

https://github.com/Saghen/blink.cmp/pull/650