feat: allow flexible configuration of the project root
Previously, the project root was always assumed to be the nearest
directory containing a `.git` directory. This commit introduces a new
`project_root_marker` option that allows users to configure how the
project root is found.

This option accepts the same arguments as the marker given to `:h
vim.fs.root()` which offers many possibilities for configuration. The
default value is still `.git`.