feat: highlight the match in the documentation window (#74)
When a match has been found in a file in the project, the documentation
window should show a preview of the match context (lines around the
match), and highlight the part where the match was found. This way the
user can quickly get an idea of where the match was found.

By default, the match is highlighted with the `BlinkRipgrepMatch`
highlight group, which is linked to (looks like) the `Search` highlight
group.