feat: add a debug mode that prints the rg command that was used
Issue
=====

It's difficult to diagnose performance issues with the plugin. Some
users have unfortunately run into this in
https://github.com/mikavilpas/blink-ripgrep.nvim/issues/82

Solution
========

Add a `debug` option to the plugin configuration that prints the rg
command that was used to get completions. The user can then copy paste
the command into their terminal and play around with it.

By default it's off, but can be enabled by setting `debug = true`.