fix: not providing completions for a project root with spaces
Issue
=====

When working in a project whose project_root_marker path contained space
characters (maybe even other special characters), the plugin would not
provide any completions.

This was caused by extra escaping of the path in the command that was
run in the background.

Solution
========

Correctly escape the spaces and provide completions.

test: can run the debug command in a shell