feat(debug,opt-in): temporarily highlight the search prefix
Issue
=====

It's hard to understand when exactly ripgrep is called to find matching
words in the project. This is an important concern because of
performance - ripgrep can take a long time to search through a large
project.

Solution
========

Temporarily highlight the search prefix when a new ripgrep search is
started. This will help users and maintainers understand the behavior of
the plugin.

To enable this feature, set `debug` to `true` in the configuration. It's
recommended to only enable this feature when debugging the plugin.