feat(opt-in): kill previous searches when a new search is started
Issue
=====

When typing text, if there are no matches for a ripgrep search, each
keypresses starts a new ripgrep search. This can lead to many searches
being active at the same time. This takes up resources and is not useful
(if the word did not match previously, it will never match with a longer
search string).

Solution
========

Stop previously running searches when a new search is started. This is
currently disabled by default (soft launch). To opt in, see the README.

If this works well for some time, it might become the default in the
future.