refactor!: a lot and add e2e tests
Issue
=====

It's not possible to not search with blink-ripgrep.nvim while
maintaining the ability to search with ripgrep (outside of Neovim).

Although ripgrep supports many ways of ignoring paths, and blink-ripgrep
is affected by those, they will always affect both ripgrep and
blink-ripgrep at the same time.

Users cannot maintain the ability to search with ripgrep in the terminal
and use blink-ripgrep in Neovim at the same time.

Solution
========

Add a new option `ignore_paths` to the configuration of blink-ripgrep.
These can be set to absolute paths (resolved by `project_root_marker`).

When a search is started, the plugin will check if the root path is
ignored, and will not search if that's the case.

BREAKING CHANGE: The `get_command` function, which provides a custom way
to generate the entire ripgrep command, must now return a table with the
command and the root path.