feat: add additional_paths option to search in additional dirs/files
Issue
=====

It's not possible to include useful reference material in the search
that exists outside the project root.

Solution
========

Add a new option `additional_paths` to the configuration. It allows
including any additional paths to search in, in addition to the project
root. This can be useful if you want to include dictionary files
(/usr/share/dict/words), framework documentation, or any other reference
material that is not available within the project root.

Solves https://github.com/mikavilpas/blink-ripgrep.nvim/issues/116