feat: enable toggling the plugin on/off with folke/snacks.nvim (#123)
Issue
=====

In some scenarios it's not nice to have blink-ripgrep pop up:
- if there are performance issues
- when giving a presentation
- when recording a video demo

Solution
========

Provide a way to toggle the plugin on/off with a keymap. Right now this
is off by default and depends on
https://github.com/folke/snacks.nvim/blob/main/docs/toggle.md

Users that don't want to use this feature or add snacks.nvim can simply
ignore it.