perf: remove unused context size from ripgrep command
The context was previously used to get the context around the search.
Getting the context has since been optimized to be done lazily when the
completion item is selected.

Currently the rg context option is unused in the command, so this might
reduce the amount of json parsing the plugin has to do.