feat: allow using gitgrep or ripgrep based on the current directory
Issue
=====

Currently, the plugin supports using ripgrep or gitgrep. It allows
opting into more performance (gitgrep), or a solution that also works
outside of git repositories.

This has to be configured before starting Neovim. However, sometimes
users need to edit files in both types of projects. This requires manual
work to keep the plugin working.

Solution
========

Add a new `gitgrep-or-ripgrep` backend. It uses gitgrep if the current
directory is a git repository, otherwise it uses ripgrep.

Right now the backend is selected based on Neovim's current working
directory (cwd). Feedback on this is welcome.