perf: avoid blocking the main thread in GitGrepOrRipgrepBackend
Issue
=====

GitGrepOrRipgrepBackend checks if the current directory is a git
repository with a shell command. If this shell command takes a long time
to execute, it will block the main thread, making the UI unresponsive.

Solution
========

Use blink.cmp.lib.async.task and avoid waiting for the shell command to
finish in a blocking manner. This allows the user to keep making edits
in neovim while the shell command is running.

It might be possible to also cancel the task cleanly, but I did not
really test this extensively.