refactor!: clearly support backend specific options (#256)
**Issue:**

Currently, two backends are supported: `ripgrep` and `gitgrep` (a
fallback backend `gitgrep-or-ripgrep` is also available). These use
ripgrep (`rg`) and `git grep` respectively to search for matches.

However, the configuration options for these backends are not clearly
separated, leading to confusion about which options apply to which
backend.

**Solution:**

Move to a new configuration format that clearly separates the
configuration options for each backend.

```lua
-- old config format
RgSource.config = {
  prefix_min_len = 3,
  fallback_to_regex_highlighting = true,
  project_root_marker = ".git",

  context_size = 5,                -- moved to backend config
  max_filesize = "1M",             -- moved to backend.ripgrep config
  additional_rg_options = {},      -- moved to backend.ripgrep config
  search_casing = "--ignore-case", -- moved to backend.ripgrep config
  ignore_paths = {},               -- moved to backend.ripgrep config
  project_root_fallback = true,    -- moved to backend.ripgrep config
  additional_paths = {},           -- moved to backend.ripgrep config

  mode = "on",
  future_features = { -- future_features has been removed
    backend = { -- this is now part of the top-level config
      use = "ripgrep",
      customize_icon_highlight = true,
    },
  },
  toggles = { on_off = nil, },
}

-- new default config
RgSource.config = {
  prefix_min_len = 3,
  fallback_to_regex_highlighting = true,
  project_root_marker = ".git",
  mode = "on",
  backend = {
    use = "ripgrep",
    customize_icon_highlight = true,
    context_size = 5,
    ripgrep = {
      ignore_paths = {},
      additional_paths = {},
      project_root_fallback = true,
      search_casing = "--ignore-case",
      max_filesize = "1M",
      additional_rg_options = {},
    },
  },
}
```

This includes an automatic migration for existing configurations that
simply moves the old options to the new format. It shows a
`vim.notify_once()` notification to inform users about the migration,
and it also shows which configuration options need to be updated.

This means that existing configurations will continue to work, but I'll
eventually remove the automatic migration, so please migrate your
configuration to the new format.