fix: git grep showing wrong part of the match as highlighted
**Issue:**

When many matches are found from the same line, `git grep` returns
incorrect columns for the matches after the first match. This causes the
highlighting to be incorrect in the documentation window.

This only happens when using the `git grep` backend, and only when
multiple matches are found on the same line.

**Solution:**

When the documentation window is being rendered, highlight all instances
of the current match manually.

For a unique match, this is the perfect workaround.

For a non-unique match, all the instances of the match are highlighted.
For example, if the match is `banana`, and there are two instances of
`banana` on the same line, both will be highlighted. This is not ideal,
but it's much better than showing the wrong part of the match as
highlighted.

Closes https://github.com/mikavilpas/blink-ripgrep.nvim/issues/228

- files specify their source backend as a field
- result files remember the prefix that started the search