feat: support multiple matches on the same line in ripgrep backend
**Issue:**

When using the ripgrep backend to match a line like the following with a
query like "ban" that matches both "banana" and "banana_with_text", only
"banana" is reported as a match, and banana_with_text is ignored
completely:

```text
banana  banana_with_text
```

Looks like I had forgotten to implement this in the ripgrep parser.

**Solution:**

Implement it! Now we report both matches as results like you would
expect.