docs: readme recommends using stable versions
Using the bleeding edge version (newest commit) can still be done
without specifying `version = "*"`, but the bleeding edge will start to
have more frequent changes due to the automerging of Dependabot PRs.

These changes don't have an effect on users (they only concern
end-to-end testing the plugin), but it might be disturbing to see
constant changes in the plugin manager.

For this reason, it's now recommended to use the latest stable version
by specifying `version = "*"`.